Update vdp_backups
I've running omd 1.20 and had issues with your version but the code was very close.  I rewrote it a bit to get it working and added in a couple some extra error handling. I removed the elapsed time because it can vary highly depending on hardware/size of VMs.  I hope this helps I appreciate your original code.